Meeting Detail
Add a basic Meeting Detail screen with a map.  The screen
is presented when the user presses a meeting (cell) from
the MeetingsViewController.  When completed, the detail
screen will show the information and map location for the
specific meeting.

Add constants files for screen size and location information.

Performed some micro improvements on existing code.  Not
changing the functionality, but improving the way the code
is written.

Changes to be committed:
	modified:   publicmeetings-ios.xcodeproj/project.pbxproj
	modified:   publicmeetings-ios/Cells/MeetingCell.swift
	modified:   publicmeetings-ios/Cells/MoreCell.swift
	new file:   publicmeetings-ios/Constants/Accuracy.swift
	new file:   publicmeetings-ios/Constants/Screen.swift
	new file:   publicmeetings-ios/Controllers/MeetingsDetailViewController.swift
	modified:   publicmeetings-ios/Controllers/MeetingsViewController.swift
	modified:   publicmeetings-ios/Controllers/MoreViewController.swift
	modified:   publicmeetings-ios/Info.plist
	new file:   publicmeetings-ios/Views/MeetingsDetailView.swift