Reminder button action
When the user presses a reminder button for a particular meeting,
toggle the setting and the image.  If the user wants to add the
reminder to the calendar, the image changes to a calendar with a
plus badge.  Otherwise, the image is a calender with no badge.

Changes to be committed:
	modified:   publicmeetings-ios/Cells/MeetingCell.swift