Meetings Badge
Add BadgeDelegate to trigger methods when a calendar reminder
button is tapped.  The button toggles whether to add the event
to your calendar and the badge will increase or decrease the
badge value accordingly.

TODO: The increment and decrement works, but logic needs to be
added to handle no reminders (hide badge) and possibly end-cases
where underflow might occur (resulting in negative values).

Additionally, the badge shouldn't appear initially, as there is
no value associated with meetings.  Currently, it serves as a
placeholder for demonstration purposes.

Changes to be committed:
	modified:   publicmeetings-ios/Cells/MeetingCell.swift
	modified:   publicmeetings-ios/Controllers/MeetingsViewController.swift
	modified:   publicmeetings-ios/TabBar/TabBarController.swift