Search view and controller
Create a search view screen and wire it up to the
search view controller.  Added dismiss keyboard
code since the keyboard is hiding the tabBar.

TODO: Need to address how this screen will look and
function.  Might add a tableview and return records
that match the search term(s).

Changes to be committed:
	modified:   publicmeetings-ios.xcodeproj/project.pbxproj
	modified:   publicmeetings-ios/Controllers/SearchViewController.swift
	new file:   publicmeetings-ios/Views/SearchView.swift