Meeting info + model
Update the model to one that is close to what we're likely to
see.  Change CreateMeetingData() to generate sample data based
on the changes.

Update the meeting cell to incorporate some of the new data.
Not a lot of time was spent formatting the date, since the
actual format will depend on the kinds of data we get from the
feed.

Changes to be committed:
	modified:   publicmeetings-ios/Cells/MeetingCell.swift
	modified:   publicmeetings-ios/Cells/MinutesCell.swift
	modified:   publicmeetings-ios/Controllers/MeetingsViewController.swift
	modified:   publicmeetings-ios/Models/Meeting.swift
	modified:   publicmeetings-ios/Utils/CreateMeetingData.swift