Meeting data + Meeting screen
Update the data model to match the data from Wichita's actual
calendar of events.

The model is currently defining everything as String for the
sake of simplicity.  The model will need to be updated using
correct data types.

Update the Meeting screen to display additional data.

With this change, we're using location to indicate where
the meeting is held.  We will need to add an additional
field to indicate whether the meeting is city, county, state,
so that the segmented control will function properly.

Changes to be committed:
	modified:   publicmeetings-ios/Cells/MeetingCell.swift
	modified:   publicmeetings-ios/Controllers/MeetingsViewController.swift
	modified:   publicmeetings-ios/Models/Meeting.swift
	modified:   publicmeetings-ios/Utils/CreateMeetingData.swift