Fixes - WebViewer
Rename pdfUrl to documentUrl so it's more clear that the viewer
will be used for differnt kinds of documents.  Also change the
methods that get called to fall in line with this change.

Change the scope for openUrl so we're passing in a parameter
instead of relying on a file-level scoped var.

Some code cleanup around spacing and extraneous lines of code.

Changes to be committed:
	modified:   publicmeetings-ios/Cells/MeetingCell.swift
	modified:   publicmeetings-ios/Controllers/AgendasViewController.swift
	modified:   publicmeetings-ios/Controllers/MinutesViewController.swift
	modified:   publicmeetings-ios/Utils/CreateMeetingData.swift
	modified:   publicmeetings-ios/WebView/WebViewer.swift