Map the meeting location
When tapping on a meeting from the meeting screen, seque
to a map page that puts a pin on the meeting location.

Known issues: The mapping works, but I have an original screen
that is being bypassed because I'm using Apple's MapKit and it's
going straight to that screen.  The seque back goes to the screen
I developed.  The mapping code I'm using is just for proof of
concept, so more work has to be done to determine how the screen
is presented.

Changes to be committed:
	modified:   publicmeetings-ios/Cells/MeetingCell.swift
	modified:   publicmeetings-ios/Controllers/MeetingsDetailViewController.swift
	modified:   publicmeetings-ios/Controllers/MeetingsViewController.swift
	modified:   publicmeetings-ios/Models/Meeting.swift