Font issue
There are instances where the font is getting truncated at the top
of labels.  The issue is due to the font being used (Dasmascus),
which was chosen at random.  Changing to the systemFont fixes the
problem, so either stick with that or look for a font that doesn't
cause truncation.

Changes to be committed:
	modified:   publicmeetings-ios/Constants/StandardValues.swift
	modified:   publicmeetings-ios/Views/AboutView.swift