Documents
Create Documents screens (view and view controller).  The Minutes
and Agenda screens are basically the same.  They only point to
different URLs.  The documents screen will combine both into one
screen (and be replaced on the tab bar).  A segmented control will
drive the data source that is displayed.

This commit serves as the first step.  Logic needs to be added to
switch from Minutes to Agendas.  Additionally, the cells for each
need to be combined into one, so that the tableView.register(_:)
is dynamic and everything from the ViewController to the Cell has
no static information (reuseIdentifier, etc).

Changes to be committed:
	modified:   publicmeetings-ios.xcodeproj/project.pbxproj
	modified:   publicmeetings-ios/Colors.xcassets/devictTan.colorset/Contents.json
	new file:   publicmeetings-ios/Controllers/DocumentsViewController.swift
	modified:   publicmeetings-ios/Controllers/MoreViewController.swift
	new file:   publicmeetings-ios/Extensions/UIApplication.swift
	modified:   publicmeetings-ios/TabBar/TabBarController.swift
	new file:   publicmeetings-ios/Views/AgendasView.swift
	new file:   publicmeetings-ios/Views/DocumentsView.swift
	new file:   publicmeetings-ios/Views/MinutesView.swift