Documents new design
Rethought the idea of the documents screen.  Instead of a segmented
control to switch between Minutes and Agenda documents, added a button
on the cell for both.  The segmented control has been removed.  The
buttons can use a bit of sprucing up.  They're just placeholders for
the time being.

Changes to be committed:
	modified:   publicmeetings-ios.xcodeproj/project.pbxproj
	new file:   publicmeetings-ios/Cells/DocumentsCell.swift
	modified:   publicmeetings-ios/Views/DocumentsView.swift