Meeting Dates
Add method that will convert mock date from String format to a Date() format and then use that to separate into its component parts and display in the DateView on the meetings cell.

Note: When we have an actual data feed, this code will be converted to handle it.

Changes to be committed:
	modified:   publicmeetings-ios/Cells/MeetingCell.swift
	modified:   publicmeetings-ios/Controllers/MeetingsViewController.swift
	modified:   publicmeetings-ios/Extensions/DateExtension.swift
	modified:   publicmeetings-ios/TabBar/TabBarController.swift
	modified:   publicmeetings-ios/Views/DateView.swift