DateView border
The initial border for DateView was askew from the cell separator.  The border color was changed to match the width and color of the separator.

Note:  The offset may be very slightly off.  Consider changing the offset of the DateView itself or somehow making a perfect alignment.

Changes to be committed:
	modified:   publicmeetings-ios/Views/DateView.swift