Background images
On the launch screen, change the image from "meetingroom" to "pillars".  Also used pillars on the About screen.

Use meetingroom for the "No Meetings" screen.  The image meetingroom is red, and the navigation bar and tab bar are blue, so having a blue image (pillars) on a mostly blue screen was a bit much.

Changes to be committed:
	modified:   publicmeetings-ios/Base.lproj/LaunchScreen.storyboard
	modified:   publicmeetings-ios/Views/AboutView.swift
	modified:   publicmeetings-ios/Views/NoMeetingsView.swift