Meeting action buttons
Change the bottomAnchor for the share and reminder buttons.  Move them closer to the bottom of the cell.  The reminder button was not in perfect visual alignment with the share button, so I moved it a bit further down, so they appear to be even with each other.

Changes to be committed:
	modified:   publicmeetings-ios/Cells/MeetingCell.swift