About screen
Update the CloseButton class.  Add a method to set the systemImage and the "regular" image.  Setting the regular image is more a courtesey and just a wrapper around the setImage(_:) method that already exists for a UIButton.

I added the setSystemImage(_:) method so that it'd be easier to change the UI for different themes.

Changes to be committed:
	modified:   publicmeetings-ios/Controllers/AboutViewController.swift
	modified:   publicmeetings-ios/Controllers/MoreViewController.swift
	modified:   publicmeetings-ios/UtilityViews/CloseButton.swift