Apple Sign In
Add the Apple sign in button from AuthenticationServices.

This commit is just to test that the authentication works.  A proper view or view controller still needs to be created and the login will need some tweaks.

The User model was also updated to accomodate the service.  The model is still not complete (waiting on the backend).  A login will only be required if the user wants to save meetings on the calendar.  Will need to provide an optional login screen.

Changes to be committed:
	modified:   publicmeetings-ios.xcodeproj/project.pbxproj
	new file:   publicmeetings-ios/Controllers/LoginViewController.swift
	modified:   publicmeetings-ios/Controllers/MoreViewController.swift
	modified:   publicmeetings-ios/Models/User.swift
	new file:   publicmeetings-ios/Public Meetings.entitlements